requirement_n.txt
asgiref
certifi
chardet
Django
django-cleanup
django-cors-headers
django-filter
django-ipware
django-js-asset
django-mptt
django-restql
djangorestframework
idna
mysqlclient
Pillow
pyPEG2
pytz
requests
sqlparse
urllib3
